Move CLI into new scai-generator package
* Remove SCAI Python implementation
* Update imports in scai-generator library
