+++ python/scai_generator/utility.py
